refactor(shared): move shared memory services to extension
